Fix imports to moved classes.
